Move the logic for TaxonomyType::persist() into a trait
